requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
watchdog==4.0.0
openpyxl==3.1.2